requirements.txt
absl-py==0.1.12
asn1crypto==0.24.0
astor==0.6.2
bleach==1.5.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.2.1
Flask==0.12.2
gast==0.2.0
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Keras==2.1.5
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.14.2
protobuf==3.5.2.post1
pycparser==2.18
pyOpenSSL==17.5.0
PySocks==1.6.8
PyYAML==3.12
requests==2.18.4
scipy==1.0.0
six==1.11.0
tensorboard==1.6.0
tensorflow-gpu==1.6.0
termcolor==1.1.0
urllib3==1.22
Werkzeug==0.14.1
win-inet-pton==1.0.1
wincertstore==0.2
